+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Recalculate function on batch allocation screen. Refs UIF-533.
 * *BREAKING* Batch allocation logs view and delete feature. Refs UIF-530.
 * Migrate to shared GA workflows. Refs UIF-572.
+* Fixes for batch allocation form. Refs UIF-579.
 
 ## [7.0.1](https://github.com/folio-org/ui-finance/tree/v7.0.1) (2024-12-02)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v7.0.0...v7.0.1)